+++ mobile/asa-go/src/components/report/AdvisoryText.tsx
@@ -6,7 +6,7 @@ import {
 } from "@/api/fbaAPI";
 import DefaultText from "@/components/report/DefaultText";
 import { useFilteredHFIStatsForDate } from "@/hooks/dataHooks";
-import { useProvincialSummaryForDate } from "@/hooks/useProvincialSummaryForDate";
+import { useProvincialSummaryForDate } from "@/hooks/dataHooks";
 import { useRunParameterForDate } from "@/hooks/useRunParameterForDate";
 import {
   getTopFuelsByArea,
